Re-fix schemas to allow decimal/real numbers in timezone ...
git-svn-id: file:///svn/phpbb/trunk@2390 89ea8834-ac86-4346-8a33-228a782c2dd0

+++ phpBB/db/schemas/mysql_schema.sql
@@ -463,7 +463,7 @@ CREATE TABLE phpbb_users (
    user_regdate int(11) DEFAULT '0' NOT NULL, 
    user_level tinyint(4) DEFAULT '0',
    user_posts mediumint(8) UNSIGNED DEFAULT '0' NOT NULL,
-   user_timezone tinyint(4) DEFAULT '0' NOT NULL,
+   user_timezone float DEFAULT '0' NOT NULL,
    user_style tinyint(4),
    user_lang varchar(255),
    user_dateformat varchar(14) DEFAULT 'd M Y H:i' NOT NULL,
